move NonEmptyList from shared_state to utils
